implement pytest-bdd instead of behave add get_ship.feature + corresponding tests made tests DRY added kener to docker compose
