+++ .devcontainer/.env
@@ -1,12 +1,13 @@
 app_port=8000
 adminer_port=8010
+db_ext_port=5432
+prometheus_port=8090
+grafana_port=8030
+kener_port=8040
+nicegui_port=8080
+alertmanager_port=8093
 db_user=star
 db_password=trek
 db_name=star-trek-db
 #db_host=startrek_db
-db_ext_port=5432
 db_url = "postgresql://star:trek@startrek_db/star-trek-db"
-prometheus_port=8090
-grafana_port=8030
-nicegui_port=8080
-alertmanager_port=8093
