+++ .devcontainer/docker-compose.yml
@@ -96,6 +96,38 @@ services:
     networks:
       articles_dc_fastapi_startrek-net:
 
+  kener:
+    image: rajnandan1/kener:latest      # Change to 'rajnandan1/kener:alpine' for an even smaller image! 😁🚀
+    container_name: kener
+    # env_file: custom.env              # Uncomment this if you are needing to export environment variables from a custom environment file. By default, Docker will import any variables that exist in `.env`
+    environment:
+      TZ: Etc/UTC
+      KENER_SECRET_KEY: replace_me_with_a_random_string   # Keep private!! - best to define in `.env` file or through Docker Secret
+      # DATABASE_URL: custom_db_url     # By default, a SQLite database is used - you may override the database url/type here
+      # RESEND_API_KEY: 
+      # RESEND_SENDER_EMAIL: 
+
+      ### You most likely will NOT need to change anything below this line. Be sure you know what you're doing!! (https://kener.ing/docs/deployment/#docker-environment-variables)
+
+      # PORT: 3000                      # Port that app listens on in the container
+      # KENER_BASE_PATH:                # By default, Kener runs at `/`. You may change this to be, e.g. `/status`, etc. Do NOT add a trailing slash!! (more info here: https://kener.ing/docs/deployment/#docker-environment-variables)
+      # ORIGIN: http://localhost:3000
+      # NODE_ENV: production            # This is already set to "production" by default within the container
+    ports:
+      - '${kener_port}:3000/tcp'
+    volumes:
+      - data:/app/database              # We suggest using a Docker named volume, which is more performant for databases
+      #- $(pwd)/uploads:/app/uploads
+      - $PWD/uploads:/app/uploads
+      #- $../kerner_uploads:/app/uploads:cached
+    # read_only: true                   # Uncommenting this fortifies security by marking the container's filesystem as read-only (aka no data can be written to the container's filesystem except for explicitly defined writable volumes and bind mounts, an exception has already been defined for `/database` and `/uploads`)
+    restart: unless-stopped
+    # depends_on:                       #  <-- Uncomment if you would like to use PostgreSQL or MySQL
+    #   - postgres                      #       ...instead of SQLite
+    #   - mysql                         #
+    networks:
+      articles_dc_fastapi_startrek-net:
+
 networks:
   articles_dc_fastapi_startrek-net:
 
@@ -106,3 +138,5 @@ volumes:
     driver: local
   alertmanager-data:
     driver: local
+  data:
+    name: kener_db
